Break down the time widget
