+++ src/Chat/Timer.tsx
@@ -1,13 +1,6 @@
 import { memo, useEffect, useState } from "react";
-
-// Font Awesome
-import Parser from "html-react-parser";
-import { library, icon } from "@fortawesome/fontawesome-svg-core";
-import { faStopwatch, faBell } from "@fortawesome/free-solid-svg-icons";
 import { TimePayload } from "../types/timePayload.type";
-library.add(faStopwatch, faBell);
-const stopwatchIcon = icon({ prefix: "fas", iconName: faStopwatch.iconName });
-const bellIcon = icon({ prefix: "fas", iconName: faBell.iconName });
+import TimeDisplay from "./TimeDisplay.tsx";
 
 const Timer = memo(function Timer(props: { serverTime: TimePayload }) {
   const [timer, setTimer] = useState(props.serverTime.timeLeftSec);
@@ -35,43 +28,10 @@ const Timer = memo(function Timer(props: { serverTime: TimePayload }) {
         }}
         aria-label="Game timer"
       >
-        <ProgressBarDisplay timer={timer} />
+        <TimeDisplay timer={timer} />
       </div>
     </div>
   );
 });
-
-interface ProgressBarDisplayProps {
-  timer: number;
-}
-
-function ProgressBarDisplay({ timer }: ProgressBarDisplayProps) {
-  if (timer > 0) {
-    return (
-      <div>
-        <>
-          {Parser(stopwatchIcon.html.toString())} {getMinutes(timer)}:
-          {getSeconds(timer)}
-        </>
-      </div>
-    );
-  } else {
-    return (
-      <span>
-        {Parser(bellIcon.html.toString())} Time&apos;s up! Who is the Spy?
-      </span>
-    );
-  }
-}
-
-function getMinutes(timer: number): string {
-  const minutes = Math.floor(timer / 60);
-  return minutes < 10 ? "0" + minutes : minutes.toString();
-}
-
-function getSeconds(timer: number): string {
-  const seconds = timer % 60;
-  return seconds < 10 ? "0" + seconds : seconds.toString();
-}
 
 export default Timer;
